Track state in a nanostore map
The typescript used in the definition of the `$cart` tells us that the
map has numbers as keys, and that the values are of type `CartItem`.

The `@ts-ignore` declaration is preventing TypeScript from whinging when
we set an item to `undefined`, which is apparently how you delete
something from a nanostore map.