Make shop-items component more reusable
We want to be able to use it on the home page, showing a limited number
of sandwiches. We don't want to be able to add them to the cart directly
from the home page either.

Note the use of `Astro.props` to access props passed into an Astro
component.