Render blog index from content collection
Note the auto-complete on the strings that are passed into
`getCollection()`, because Astro has sorted out the types for us when
generating the `astro:content` module from the `defineCollection()`
call in the previous commit.

We get completion on all the `post.data` properties too, which are the
fields in the Markdown frontmatter.

If we moved the posts off the filesystem and into a headless CMS, all
we'd need to do is replace the call to `getCollection()` with something
that loads it from the CMS, passing it through the same Zod schema. It'd
just work.