Render each blog post at build time
We're using the `getStaticPaths` function to return an array of objects
that describe each page. That list of objects tells Astro about all of the
possible statically rendered pages that can be served by this component.

Note that when we instantiate `post` by reading `Astro.props` that we're
getting the post object for a specific slug.

The `<Content />` component is created for us from the body of the
associated Markdown file.