Display message on newsletter subscription
Note that this page works perfectly with JavaScript disabled; it's just
using classic HTTP forms and server-side rendering.

The `Astro.request` object is the standard Request object, and
formData() returns the standard FormData object.

Note that we can conditionally include a CSS class with `class:list`.
Very handy. Also note that we can exclude a class from the list by using
`!` to negate when we apply the class; the value is just an expression.